requirements.txt
Flask
Flask-SQLAlchemy
Flask-JWT-Extended
Flasgger
Flask-Cors
python-dotenv
